fix linux32 builder tests

+++ rts/System/Platform/Misc.cpp
@@ -401,9 +401,9 @@ namespace Platform
 			return -1;
 
 		if (st.f_frsize != 0)
-			return ((st.f_frsize * st.f_bavail) / (1024 * 1024));
+			return (((uint64_t)st.f_frsize * st.f_bavail) / (1024 * 1024));
 
-		return ((st.f_bsize * st.f_bavail) / (1024 * 1024));
+		return (((uint64_t)st.f_bsize * st.f_bavail) / (1024 * 1024));
 		#endif
 	}
